requirements.txt
streamlit==1.32.0
beautifulsoup4==4.12.3
bs4
requests==2.31.0
transformers==4.38.2
torch==2.2.2
gtts==2.5.1
pandas
nltk
fastapi
uvicorn
selenium==4.18.1
webdriver-manager
plotly==5.24.1
googletrans==4.0.0-rc1
ffmpeg-python
imageio[ffmpeg]
apscheduler




# Compatibility-safe HTTP libraries

httpx==0.13.3  # Pinned for googletrans compatibility
httpcore==0.9.1  # Pinned for httpx 0.13.3 compatibility
